Completed basic example instrumenter.
The example instrumenter now includes transformations to add the requisite
include files as well as translating check macro arguments properly and
instrumenting source file main routines with calls to the runtime's new
initialization and finalization routines.

ROSE-related output directories are now consistent between printer and
transformation examples.

A transformation Makefile was added for the ContractsInstrumenter example
for building, executing, and performing a very rudimentary check (just a
comparison) of the outputs from the transformed codes.

It is important to note that, at this time, there is still no connection
between the actual assertion expressions and the resulting instrumentation.
This is expected to be dealt with during my next project window.