Addressed ASSERT instrumentation and short-circuited clause processing.
Corrected the ordering of instrumented ASSERT checks when there are
multiple assertion expressions in the "clause".

Also reduced visitor processing of contract comments by limiting it
to the "first" time a line is processed (since that is the node
with the associated comments).

Unfortunately, have not been able to determine why the helloworld-v2.cpp
file's FINAL comment (associated with a "return 0;" statement) does not
appear to be handed to the visitor.