Snapshot of latest contracts instrumentation work -- mostly testing.
Was able to determine my installation of ROSE does NOT provide visit()
with an AST node for the 'return 0;' associated with the in-lined
'CONTRACT FINAL' annotation in the helloworld-v2.cc example.  Consequently,
the corresponding finalization call is not instrumented.  Added
helloworld-v3.cc example to illustrate that a simple print to illustrate
a case where a node with the associated contract annotation does result
in the finialization call being instrumented as desired.

Summary of Changes:
- Added helloworld-v3.cc (for illustrating inlined CONTRACT FINAL);
- Added oracle-based test cases for instrumentation of example;
- Completed support for overall (examples) test suite status reporting
  (when using 'make check');
- Revised assorted error message formatting.