Generating selected built-in expression support and return type declarations.
Added necessary generation of the return type declaration and the include file
for the (two) built-in expressions (i.e., pce_in_range and pce_near_equal).
At present only two examples use a built-in expression (and only pce_in_range),
but testing reveals the code appears to be properly generated and the
instrumented codes build and execute properly.